requirements.txt
flask
python-binance
gspread
oauth2client
